Se crearon nuevos permisos

+++ app/Http/Controllers/Admin/AnalisisdeRiesgosController.php
@@ -9,6 +9,8 @@ use App\Models\Empleado;
 use Illuminate\Http\Request;
 use Laracasts\Flash\Flash;
 use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class AnalisisdeRiesgosController extends Controller
 {
@@ -24,6 +26,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function index(Request $request)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->ajax()) {
             //Esta es el error , activo_id no lo encuentra, hay que modificar la relacion en el modelo de matrizriesgo
             $query = AnalisisDeRiesgo::orderByDesc('id')->get();
@@ -33,9 +36,9 @@ class AnalisisdeRiesgosController extends Controller
             $table->addColumn('actions', '&nbsp;');
 
             $table->editColumn('actions', function ($row) {
-                $viewGate = 'user_show';
-                $editGate = 'user_edit';
-                $deleteGate = 'user_delete';
+                $viewGate = 'analisis_de_riesgos_matriz_riesgo_show';
+                $editGate = 'analisis_de_riesgos_matriz_riesgo_edit';
+                $deleteGate = 'analisis_de_riesgos_matriz_riesgo_delete';
                 $crudRoutePart = 'analisis-riesgos';
 
                 return view('partials.datatablesActions', compact(
@@ -96,6 +99,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $empleados = Empleado::get();
 
         //$tipoactivos = Tipoactivo::all()->pluck('tipo', 'id')->prepend(trans('global.pleaseSelect'), '');
@@ -111,6 +115,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $analisis = AnalisisDeRiesgo::create($request->all());
         switch ($request->tipo) {
             case 'Seguridad de la información':
@@ -133,6 +138,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function show(Request $request, $id)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $analisis = AnalisisDeRiesgo::find($id);
 
         return view('admin.analisis-riesgos.show', compact('analisis'));
@@ -146,6 +152,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function edit($id)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $empleados = Empleado::get();
         $analisis = AnalisisDeRiesgo::find($id);
 
@@ -161,6 +168,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function update(Request $request, $id)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $analisis = AnalisisDeRiesgo::find($id);
 
         $analisis->update([
@@ -183,6 +191,7 @@ class AnalisisdeRiesgosController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(Gate::denies('analisis_de_riesgos_matriz_riesgo_destroy'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $analisis = AnalisisDeRiesgo::find($id);
         $analisis->delete();
