+++ app/Http/Controllers/Admin/PerfilController.php
@@ -7,11 +7,14 @@ use App\Models\Empleado;
 use App\Models\PerfilEmpleado;
 use Illuminate\Http\Request;
 use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Http\Response;
 
 class PerfilController extends Controller
 {
     public function index(Request $request)
     {
+        abort_if('niveles_jerarquicos_access', Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->ajax()) {
             $query = PerfilEmpleado::with(['empleados'])->orderBy('id')->get();
             $table = DataTables::of($query);
@@ -20,9 +23,9 @@ class PerfilController extends Controller
             $table->addColumn('actions', '&nbsp;');
 
             $table->editColumn('actions', function ($row) {
-                $viewGate = 'user_show';
-                $editGate = 'user_edit';
-                $deleteGate = 'user_delete';
+                $viewGate = 'niveles_jerarquicos_create';
+                $editGate = 'niveles_jerarquicos_edit';
+                $deleteGate = 'niveles_jerarquicos_delete';
                 $crudRoutePart = 'perfiles';
 
                 return view('partials.datatablesActions', compact(
@@ -54,6 +57,7 @@ class PerfilController extends Controller
 
     public function create()
     {
+        abort_if('niveles_jerarquicos_create', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $empleados = Empleado::get();
 
         return view('admin.perfiles.create', compact('empleados'));
@@ -61,6 +65,8 @@ class PerfilController extends Controller
 
     public function store(Request $request)
     {
+        abort_if('niveles_jerarquicos_create', Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $perfil = PerfilEmpleado::create($request->all());
 
         return redirect()->route('admin.perfiles.index')->with('success', 'Guardado con éxito');
@@ -68,6 +74,7 @@ class PerfilController extends Controller
 
     public function edit($perfil)
     {
+        abort_if('niveles_jerarquicos_edit', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
 
         return view('admin.perfiles.edit', compact('perfil'));
@@ -75,6 +82,7 @@ class PerfilController extends Controller
 
     public function update(Request $request, $perfil)
     {
+        abort_if('niveles_jerarquicos_edit', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
         $perfil->update($request->all());
 
@@ -83,6 +91,7 @@ class PerfilController extends Controller
 
     public function show($perfil)
     {
+        abort_if('niveles_jerarquicos_show', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
 
         return view('admin.perfiles.show', compact('perfil'));
@@ -90,6 +99,7 @@ class PerfilController extends Controller
 
     public function destroy($perfil)
     {
+        abort_if('niveles_jerarquicos_delete', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
         // dd($perfil);
         $perfil->delete();
