+++ app/Http/Controllers/Admin/RH/EV360CompetenciasController.php
@@ -10,13 +10,16 @@ use App\Models\RH\CompetenciaPuesto;
 use App\Models\RH\Conducta;
 use App\Models\RH\EvaluacionRepuesta;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class EV360CompetenciasController extends Controller
 {
     public function index(Request $request)
     {
+        // abort_if(Gate::denies('capital_humano_competencias_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->ajax()) {
             $competencias = Competencia::with('tipo')->get();
 
@@ -28,6 +31,7 @@ class EV360CompetenciasController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('capital_humano_competencias_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $competencia = new Competencia;
         $tipo_seleccionado = null;
 
@@ -36,6 +40,7 @@ class EV360CompetenciasController extends Controller
 
     public function store(Request $request)
     {
+        abort_if(Gate::denies('capital_humano_competencias_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $request->validate([
             'nombre' => 'required|string|max:255',
             'descripcion' => 'nullable|string',
@@ -67,6 +72,8 @@ class EV360CompetenciasController extends Controller
 
     public function storeAndRedirect(Request $request)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $request->validate([
             'nombre' => 'required|string|max:255',
             'descripcion' => 'nullable|string',
@@ -82,6 +89,8 @@ class EV360CompetenciasController extends Controller
 
     public function edit($competencia, $onlyConductas = null)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $competencia = Competencia::find(intval($competencia));
         $tipo_seleccionado = $competencia->tipo_id;
         $editar_solo_conductas = false;
@@ -94,6 +103,8 @@ class EV360CompetenciasController extends Controller
 
     public function update(Request $request, $competencia)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $request->validate([
             'nombre' => 'required|string|max:255',
             'descripcion' => 'nullable|string',
@@ -142,6 +153,8 @@ class EV360CompetenciasController extends Controller
 
     public function conductas(Request $request, $competencia)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->ajax()) {
             // $conductas = Conducta::where('competencia_id', intval($competencia))->get()->sortBy('ponderacion');
             $conductas = Conducta::where('competencia_id', intval($competencia))->orderBy('ponderacion')->get();
@@ -152,6 +165,8 @@ class EV360CompetenciasController extends Controller
 
     public function show($id)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $competencia = Competencia::find($id);
 
         return view('admin.recursos-humanos.evaluacion-360.competencias.show', compact('competencia'));
@@ -223,9 +238,11 @@ class EV360CompetenciasController extends Controller
 
     public function destroy(Competencia $competencia)
     {
+
+        abort_if(Gate::denies('capital_humano_competencias_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         // $competencia = Competencia::find($competencia);
         $competencia->delete();
 
-        return response()->json(['deleted'=> true]);
+        return response()->json(['deleted' => true]);
     }
 }
